# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Forth Architecture Consulting & Construction Ltd API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorEnvelope {
  error: string;
}

export interface Project {
  id: number;
  title: string;
  description: string;
  /** @nullable */
  coverImage: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  completionDate?: string | null;
  category: string;
  featured: boolean;
  sortOrder?: number;
  createdAt: string;
}

export interface ProjectImage {
  id: number;
  projectId: number;
  imageUrl: string;
  /** @nullable */
  caption?: string | null;
  sortOrder?: number;
  createdAt: string;
}

export interface ProjectWithImages {
  id: number;
  title: string;
  description: string;
  /** @nullable */
  coverImage: string | null;
  /** @nullable */
  location?: string | null;
  /** @nullable */
  completionDate?: string | null;
  category: string;
  featured: boolean;
  sortOrder?: number;
  createdAt: string;
  images: ProjectImage[];
}

export interface ProjectList {
  data: Project[];
  total: number;
  page: number;
  limit: number;
}

export interface ProjectInput {
  /** @minLength 1 */
  title: string;
  description: string;
  coverImage?: string;
  location?: string;
  completionDate?: string;
  category: string;
  featured?: boolean;
  sortOrder?: number;
}

export interface ProjectUpdate {
  /** @minLength 1 */
  title?: string;
  description?: string;
  coverImage?: string;
  location?: string;
  completionDate?: string;
  category?: string;
  featured?: boolean;
  sortOrder?: number;
}

export interface ProjectImageInput {
  imageUrl: string;
  caption?: string;
  sortOrder?: number;
}

export interface GalleryItem {
  id: number;
  title: string;
  imageUrl: string;
  category: string;
  /** @nullable */
  description?: string | null;
  sortOrder?: number;
  createdAt: string;
}

export interface GalleryItemInput {
  /** @minLength 1 */
  title: string;
  imageUrl: string;
  category: string;
  description?: string;
  sortOrder?: number;
}

export interface GalleryItemUpdate {
  title?: string;
  imageUrl?: string;
  category?: string;
  description?: string;
  sortOrder?: number;
}

export interface TeamMember {
  id: number;
  name: string;
  position: string;
  /** @nullable */
  photo?: string | null;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  linkedinUrl?: string | null;
  /** @nullable */
  twitterUrl?: string | null;
  sortOrder?: number;
  createdAt: string;
}

export interface TeamMemberInput {
  /** @minLength 1 */
  name: string;
  position: string;
  photo?: string;
  bio?: string;
  email?: string;
  linkedinUrl?: string;
  twitterUrl?: string;
  sortOrder?: number;
}

export interface TeamMemberUpdate {
  name?: string;
  position?: string;
  photo?: string;
  bio?: string;
  email?: string;
  linkedinUrl?: string;
  twitterUrl?: string;
  sortOrder?: number;
}

export interface Service {
  id: number;
  title: string;
  description: string;
  icon: string;
  /** @nullable */
  imageUrl?: string | null;
  sortOrder?: number;
  createdAt: string;
}

export interface ServiceInput {
  /** @minLength 1 */
  title: string;
  description: string;
  icon: string;
  imageUrl?: string;
  sortOrder?: number;
}

export interface ServiceUpdate {
  title?: string;
  description?: string;
  icon?: string;
  imageUrl?: string;
  sortOrder?: number;
}

export interface SocialLink {
  id: number;
  platform: string;
  url: string;
  createdAt: string;
}

export interface SocialLinkInput {
  /** @minLength 1 */
  platform: string;
  /** @minLength 1 */
  url: string;
}

export interface SocialLinkUpdate {
  platform?: string;
  url?: string;
}

export interface CompanyInfo {
  id: number;
  companyName: string;
  /** @nullable */
  slogan?: string | null;
  /** @nullable */
  about?: string | null;
  /** @nullable */
  mission?: string | null;
  /** @nullable */
  vision?: string | null;
  /** @nullable */
  values?: string | null;
  /** @nullable */
  phone?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  whatsapp?: string | null;
  /** @nullable */
  logoUrl?: string | null;
  /** @nullable */
  mapEmbedUrl?: string | null;
}

export interface CompanyInfoUpdate {
  companyName?: string;
  slogan?: string;
  about?: string;
  mission?: string;
  vision?: string;
  values?: string;
  phone?: string;
  email?: string;
  address?: string;
  whatsapp?: string;
  logoUrl?: string;
  mapEmbedUrl?: string;
}

export interface ContactSubmission {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  email: string;
  phone?: string;
  /** @minLength 1 */
  subject: string;
  /** @minLength 1 */
  message: string;
}

export interface ContactMessage {
  id: number;
  name: string;
  email: string;
  /** @nullable */
  phone?: string | null;
  subject: string;
  message: string;
  isRead: boolean;
  createdAt: string;
}

export type DashboardStatsProjectsByCategoryItem = {
  category: string;
  count: number;
};

export type DashboardStatsGalleryByCategoryItem = {
  category: string;
  count: number;
};

export interface DashboardStats {
  projectCount: number;
  teamCount: number;
  galleryCount: number;
  serviceCount: number;
  unreadMessages: number;
  featuredProjects: number;
  projectsByCategory?: DashboardStatsProjectsByCategoryItem[];
  galleryByCategory?: DashboardStatsGalleryByCategoryItem[];
}

export interface UploadUrlRequest {
  /** @minLength 1 */
  name: string;
  /** @minimum 1 */
  size: number;
  /** @minLength 1 */
  contentType: string;
}

export interface UploadUrlResponse {
  uploadURL: string;
  objectPath: string;
  metadata?: UploadUrlRequest;
}

export type ListProjectsParams = {
category?: string;
search?: string;
/**
 * @minimum 1
 */
page?: number;
/**
 * @minimum 1
 * @maximum 100
 */
limit?: number;
featured?: boolean;
};

export type ListGalleryParams = {
category?: string;
};

export type ListContactMessagesParams = {
read?: boolean;
};