# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Forth Architecture Consulting & Construction Ltd API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 12
        - name: featured
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Projects list with pagination
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectList"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a project by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectWithImages"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /projects/{id}/images:
    get:
      operationId: listProjectImages
      tags: [projects]
      summary: List images for a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project images
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectImage"
    post:
      operationId: addProjectImage
      tags: [projects]
      summary: Add an image to a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectImageInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectImage"
  /project-images/{imageId}:
    delete:
      operationId: deleteProjectImage
      tags: [projects]
      summary: Delete a project image
      parameters:
        - name: imageId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  # ── Gallery ──────────────────────────────────────────────────────────────
  /gallery:
    get:
      operationId: listGallery
      tags: [gallery]
      summary: List gallery items
      parameters:
        - name: category
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Gallery items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GalleryItem"
    post:
      operationId: createGalleryItem
      tags: [gallery]
      summary: Create a gallery item
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GalleryItemInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GalleryItem"
  /gallery/{id}:
    get:
      operationId: getGalleryItem
      tags: [gallery]
      summary: Get a gallery item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Gallery item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GalleryItem"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    patch:
      operationId: updateGalleryItem
      tags: [gallery]
      summary: Update a gallery item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GalleryItemUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GalleryItem"
    delete:
      operationId: deleteGalleryItem
      tags: [gallery]
      summary: Delete a gallery item
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Team Members ─────────────────────────────────────────────────────────
...
  /services:
    get:
      operationId: listServices
      tags: [services]
      summary: List services
      responses:
        "200":
          description: Services
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Service"
    post:
      operationId: createService
      tags: [services]
      summary: Create a service
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
  /services/{id}:
    patch:
      operationId: updateService
      tags: [services]
      summary: Update a service
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServiceUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"
    delete:
      operationId: deleteService
      tags: [services]
      summary: Delete a service
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Social Links ──────────────────────────────────────────────────────────
...
  /contact:
    post:
      operationId: submitContact
      tags: [contact]
      summary: Submit a contact form
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactSubmission"
      responses:
        "201":
          description: Contact submission recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactMessage"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /contact/messages:
    get:
      operationId: listContactMessages
      tags: [contact]
      summary: List contact messages (admin)
      parameters:
        - name: read
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Contact messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ContactMessage"
  /contact/messages/{id}/read:
    patch:
      operationId: markMessageRead
      tags: [contact]
      summary: Mark a contact message as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactMessage"

  # ── Stats (wow endpoint) ──────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # ── Project schemas ─────────────────────────────────────────────────────
    Project:
      type: object
      required: [id, title, description, coverImage, category, featured, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        coverImage:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        completionDate:
          type: ["string", "null"]
        category:
          type: string
        featured:
          type: boolean
        sortOrder:
          type: integer
        createdAt:
          type: string
    ProjectWithImages:
      type: object
      required: [id, title, description, coverImage, category, featured, createdAt, images]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        coverImage:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        completionDate:
          type: ["string", "null"]
        category:
          type: string
        featured:
          type: boolean
        sortOrder:
          type: integer
        createdAt:
          type: string
        images:
          type: array
          items:
            $ref: "#/components/schemas/ProjectImage"
    ProjectList:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Project"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    ProjectInput:
      type: object
      required: [title, description, category]
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        coverImage:
          type: string
        location:
          type: string
        completionDate:
          type: string
        category:
          type: string
        featured:
          type: boolean
        sortOrder:
          type: integer
    ProjectUpdate:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        coverImage:
          type: string
        location:
          type: string
        completionDate:
          type: string
        category:
          type: string
        featured:
          type: boolean
        sortOrder:
          type: integer
    ProjectImage:
      type: object
      required: [id, projectId, imageUrl, createdAt]
      properties:
        id:
          type: integer
        projectId:
          type: integer
        imageUrl:
          type: string
        caption:
          type: ["string", "null"]
        sortOrder:
          type: integer
        createdAt:
          type: string
    ProjectImageInput:
      type: object
      required: [imageUrl]
      properties:
        imageUrl:
          type: string
        caption:
          type: string
        sortOrder:
          type: integer

    # ── Gallery schemas ────────────────────────────────────────────────────
    GalleryItem:
      type: object
      required: [id, title, imageUrl, category, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        imageUrl:
          type: string
        category:
          type: string
        description:
          type: ["string", "null"]
        sortOrder:
          type: integer
        createdAt:
          type: string
    GalleryItemInput:
      type: object
      required: [title, imageUrl, category]
      properties:
        title:
          type: string
          minLength: 1
        imageUrl:
          type: string
        category:
          type: string
        description:
          type: string
        sortOrder:
          type: integer
    GalleryItemUpdate:
      type: object
      properties:
        title:
          type: string
        imageUrl:
          type: string
        category:
          type: string
        description:
          type: string
        sortOrder:
          type: integer

    # ── Team schemas ───────────────────────────────────────────────────────
...
    Service:
      type: object
      required: [id, title, description, icon, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        icon:
          type: string
        imageUrl:
          type: ["string", "null"]
        sortOrder:
          type: integer
        createdAt:
          type: string
    ServiceInput:
      type: object
      required: [title, description, icon]
      properties:
        title:
          type: string
          minLength: 1
        description:
          type: string
        icon:
          type: string
        imageUrl:
          type: string
        sortOrder:
          type: integer
    ServiceUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        icon:
          type: string
        imageUrl:
          type: string
        sortOrder:
          type: integer

    # ── Social Link schemas ────────────────────────────────────────────────
...
    ContactSubmission:
      type: object
      required: [name, email, subject, message]
      properties:
        name:
          type: string
          minLength: 1
        email:
          type: string
          minLength: 1
        phone:
          type: string
        subject:
          type: string
          minLength: 1
        message:
          type: string
          minLength: 1
    ContactMessage:
      type: object
      required: [id, name, email, subject, message, isRead, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: ["string", "null"]
        subject:
          type: string
        message:
          type: string
        isRead:
          type: boolean
        createdAt:
          type: string

    # ── Stats schema ───────────────────────────────────────────────────────
...